TCA-54 the MainActivity and its viewmodel were relocated
