+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:theme="@style/Theme.TwitterCloneAppWithActionBar">
         </activity>
     </application>
