+++ app/src/main/java/ar/scacchipa/twittercloneapp/presentation/splash/FragmentSplash.kt
@@ -9,7 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import ar.scacchipa.twittercloneapp.R
 import ar.scacchipa.twittercloneapp.databinding.FragmentSplashLayoutBinding
-import ar.scacchipa.twittercloneapp.presentation.MainActivity
+import ar.scacchipa.twittercloneapp.presentation.main.MainActivity
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class FragmentSplash: Fragment() {
